+++ CHANGELOG.md
@@ -10,8 +10,9 @@ and this project adheres to
 
 ### Added
 
-- ✨(fullstack)  allow participants to mute others based on room configuration
+- ✨(fullstack) allow participants to mute others based on room configuration
 - ✨(frontend) add synchronizer for room metadata updates
+- ✨(frontend) make reaction toolbar responsive on small viewports
 
 ### Changed