+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - ✨(fullstack) allow participants to mute others based on room configuration
 - ✨(frontend) add synchronizer for room metadata updates
 - ✨(frontend) make reaction toolbar responsive on small viewports
+- ✨(frontend) enable reactions on mobile devices
 
 ### Changed