+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to
 - ✨(backend) expose room configuration to all API consumers
 - 🩹(frontend) improve reaction toolbar centering with dynamic positioning
 - 🚀 (paas) remove buildpack requirements.txt to use the new uv.lock #1349
+- ✨(backend) allow room configuration and access level via external api #1260
 
 ## [1.16.0] - 2026-05-13